+++ src/routes/me/index.ts
@@ -1,11 +1,10 @@
 import status from "./status";
 import rename from "./rename";
-import block from "./block";
-import unblock from "./unblock";
 import blocked from "./blocked";
 import avatar from "./avatar";
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
 import votes from "./votes";
+import starred from "./starred";
 
 export default function (
     fastify: FastifyInstance,
@@ -14,10 +13,9 @@ export default function (
 ) {
     fastify.register(status);
     fastify.register(rename);
-    fastify.register(block);
-    fastify.register(unblock);
-    fastify.register(blocked);
     fastify.register(avatar);
     fastify.register(votes);
+    fastify.register(blocked);
+    fastify.register(starred);
     done();
 }