+++ src/routes/user/index.ts
@@ -3,6 +3,8 @@ import avatar from "./avatar";
 import threads from "./threads";
 import name from "./name";
 import profile from "./profile";
+import block from "./actions/block";
+import unblock from "./actions/unblock";
 
 export default function (
     fastify: FastifyInstance,
@@ -13,5 +15,7 @@ export default function (
     fastify.register(name);
     fastify.register(avatar);
     fastify.register(threads);
+    fastify.register(block);
+    fastify.register(unblock);
     done();
 }